+++ CHANGELOG.md
@@ -27,6 +27,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Option to disable video recording of sessions
+
 ### Changed
 
 ### Deprecated